Implemented Stamina!
Just added in a stamina system! First question out of the way, no this is not because of botw.
I originally wanted a system that would tie all the actions of users across races, be it galloping, flying or magic, all under one universal indicator.
The good part is calling the system 'stamina' makes a lot of sense! The bad part is then I remembered botw exists and has a pattern for the UI so optimal that it was going to be hard to diverge from it without getting crappier.
I settled for a less optimal type of UI in order to avoid some shape conflicts but the colors were far harder to escape from, so they're the same.

Anyways:

- Added the stamina system and UI! This *should* be generic enough a meter to be dealt with properly for all the use cases I could probably come up with. It's tied into galloping, flying, *and* magic, all now under one cozy umbrella.

- Added sound and VFX to indicate when magic is unusable (i.e. while tired).

- Updated the color of sweating to be more easily visible, as well as added a back version so feedback for being tired isn't accidentally obscured.

- Added a poof of dust when you begin galloping or abruptly change directions during a gallop!

- Noticed a glitch where a 100% siphoned/drained stamina meter allows pegasi to "glide" because of the leniency for flapping while tired and how fast the gauge "refills" while "completely drained."
  I've decided to keep it because I personally really like it and it's more or less benign because all you can get is horizontal distance, not vertical height, from it. Maybe it could be useful or I could make some cheeky secrets on the map that call for exploiting it but it's so normal (and teleporting as a unicorn or just flying normally could do just as well and literally better) that I don't think very much mileage will be gotten from it. It'd be pretty hilarious to be proved wrong though.